Move macOS dSYM search code to separate file
